requirements.txt
asgiref==3.5.2
Django==4.0.4
django-crum==0.7.9
django-widget-tweaks==1.4.12
Pillow==9.1.1
sqlparse==0.4.2
